editoast: views: move `project` in `operational_studies::hierarchy`
