+++ editoast/src/views/operational_studies/hierarchy.rs
@@ -1,2 +1,3 @@
+pub mod project;
 pub mod scenario;
 pub mod study;
