+++ editoast/src/views/operational_studies/hierarchy/study.rs
@@ -18,13 +18,13 @@ use thiserror::Error;
 use utoipa::IntoParams;
 use utoipa::ToSchema;
 
+use super::project::ProjectError;
 use crate::error::InternalError;
 use crate::error::Result;
 use crate::views::operational_studies::OperationalStudiesOrderingParam;
 use crate::views::pagination::PaginatedList as _;
 use crate::views::pagination::PaginationQueryParams;
 use crate::views::pagination::PaginationStats;
-use crate::views::project::ProjectError;
 use editoast_models::prelude::*;
 use editoast_models::project::Error as ProjectModelError;
 use editoast_models::project::Project;
